% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycalc.R
\name{mycalc}
\alias{mycalc}
\title{A vector arithmetic function}
\usage{
mycalc(x, y, fun)
}
\arguments{
\item{x}{A vector or matrice}

\item{y}{A vector or matrice}
}
\description{
This is a function which takes three arguments, “x”, “y”, and “fun”,
which stands for “function”. Arguments “x” and “y” are vectors
or matrices, while argument “fun” will has four possible values:
“sum”, “subtract”, “divide”, “multiply”.Function mycalc will use conditional
statements to choose the correct calculation to make based on the value you
argument value given to operator.
}
\examples{
mycalc(x=x,y=y,"sum")
mycalc(x=x,y=y,"subtract")
mycalc(x=x,y=y,"divide")
mycalc(x=x,y=y,"multiply")
}
